move interfacebase objects into rcpp_interface_base instead of individual interface module files
